+++ shared-libs/src/main/java/shared/libs/dto/ProductDto.java
@@ -13,7 +13,7 @@ public class ProductDto {
     private Integer id;
     private String name;
     private ProductTypeDto type;
-    private Integer amount;
+    private Double amount;
     private Double price;
     private DiscountDto discount;
 
@@ -21,6 +21,6 @@ public class ProductDto {
         return String.format("Id: %d\n" +
                 "Name_product: %s\n" +
                 "Price: %.2f\n" +
-                "Amount: %d", getId(), getName(), getPrice(), getAmount());
+                "Amount: %.2f", getId(), getName(), getPrice(), getAmount());
     }
 }
